ensure we expect the correct amount of assertions
It is a bit of noise but it will ensure that we don't have false
positives. For example when you have the following code:

```js
it('should fail', () => {
  return promise.catch((err) => {
    expect(err).toBeInstanceOf(Error)
  });
});
```

You would expect that the test passes because we have an expect in the
`.catch` block. However if the promise just resolves it means that no
assertions are executed and the test is considered successful. So in
this case it will _also_ call but for the incorrect reasons!